requirements.txt
certifi==2020.12.5
chardet==4.0.0
elasticsearch==7.12.1
elasticsearch-dsl==7.3.0
future==0.18.2
idna==2.10
kafka-python==2.0.2
numpy==1.20.2
oauthlib==3.1.0
pandas==1.2.4
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.0
python-twitter==3.5
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.10.0
urllib3==1.26.4
